Metavus 1.2.1

+++ plugins/Captcha/lib/securimage/WavFile.php
@@ -1701,7 +1701,10 @@ class WavFile
 
                 /************* DEGRADE FILTER *******************/
                 if ($filter_degrade) {
-                    $sampleFloat += rand(1000000 * ($degrade_quality - 1), 1000000 * (1 - $degrade_quality)) / 1000000;
+                    $sampleFloat += rand(
+                        (int)(1000000 * ($degrade_quality - 1)),
+                        (int)(1000000 * (1 - $degrade_quality))
+                    ) / 1000000;
                 }
 
                 /************* VOLUME FILTER *******************/
